extract total column
